Fix : 생성자 대신 builder 사용, FeedDTO 통합
