+++ src/main/java/com/grinder/domain/entity/Feed.java
@@ -1,6 +1,6 @@
 package com.grinder.domain.entity;
 
-import com.grinder.domain.dto.FeedRequestDTO;
+import com.grinder.domain.dto.FeedDTO;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -46,12 +46,6 @@ public class Feed extends BaseEntity {
         feedId = feedId == null ? UUID.randomUUID().toString() : feedId;
         hits = hits == null ? 0 : hits;
         isVisible = isVisible == null ? true : isVisible;
-    }
-
-    public Feed(String feedId, FeedRequestDTO request, Member member, Cafe cafe) {
-        this.feedId = feedId;
-        this.member = member;
-        this.cafe = cafe;
     }
 
     public void updateFeed(Cafe cafe, String content, Integer grade) {
