+++ src/main/java/com/grinder/domain/entity/Image.java
@@ -34,11 +34,5 @@ public class Image {
     public void prePersist() {
         imageId = imageId == null ? UUID.randomUUID().toString() : imageId;
     }
-
-    public Image(String imageUrl, String contentId, ContentType contentType) {
-        this.imageUrl = imageUrl;
-        this.contentId = contentId;
-        this.contentType = contentType;
-    }
 }
