+++ src/main/java/com/grinder/domain/entity/Tag.java
@@ -32,9 +32,4 @@ public class Tag {
     public void prePersist() {
         tagId = tagId == null ? UUID.randomUUID().toString() : tagId;
     }
-
-    public Tag(Feed feed, TagName tagName) {
-        this.feed = feed;
-        this.tagName = tagName;
-    }
 }
